apps/video-engine/pyproject.toml
[project]
name = "video-engine"
version = "0.1.0"
description = "Video Engine service for Auto-Short-Factory using FFmpeg"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=10.1.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
